+++ src/App.js
@@ -1,56 +1,49 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
+import MainPage from "./pages/mainPage/MainPage";
 
 function App() {
   return (
     <Router>
-      <Header />
-      <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/about" element={<About />} />
-      </Routes>
+      {/*<Header />*/}
+      {/*    <Sidebar />*/}
+          <Routes>
+            <Route path="/" element={<MainPage />} />
+          </Routes>
     </Router>
   );
 }
 
-function Header() {
-  return (
-    <header style={headerStyle}>
-      <Link to="/" style={{ textDecoration: 'none' }}>
-        <img
-          src="/logo.svg"
-          alt="YouTube Logo"
-          style={logoStyle}
-        />
-      </Link>
-    </header>
-  );
-}
-
-function Home() {
-  return <h1 style={pageStyle}>Welcome to Home</h1>;
-}
-
-function About() {
-  return <h1 style={pageStyle}>About Page</h1>;
-}
-
-const headerStyle = {
-  display: 'flex',
-  alignItems: 'center',
-  padding: '10px 20px',
-  backgroundColor: '#ffffff',
-  color: 'white',
-};
-
-const logoStyle = {
-  height: '40px',
-  cursor: 'pointer',
-};
-
-const pageStyle = {
-  textAlign: 'center',
-  marginTop: '50px',
-};
+// function Header() {
+//   return (
+//     <header style={headerStyle}>
+//       <Link to="/" style={{ textDecoration: 'none' }}>
+//         <img
+//           src="/logo.svg"
+//           alt="YouTube Logo"
+//           style={logoStyle}
+//         />
+//       </Link>
+//     </header>
+//   );
+// }
+//
+// const headerStyle = {
+//   display: 'flex',
+//   alignItems: 'center',
+//   padding: '10px 20px',
+//   backgroundColor: '#ffffff',
+//   color: 'white',
+// };
+//
+// const logoStyle = {
+//   height: '40px',
+//   cursor: 'pointer',
+// };
+//
+// const pageStyle = {
+//   textAlign: 'center',
+//   marginTop: '50px',
+// };
 
 export default App;
