+++ src/pages/mainPage/MainPage.css
@@ -1,0 +1,4 @@
+.pageStyle {
+    textAlign: center;
+    size: 200px;
+}
